Minor changes: Fixed font-sizing of App title, created global constants as enum for About and Footer components
