+++ ecommerce/ecommerce/Constants/GlobalConstants.swift
@@ -1,0 +1,24 @@
+//
+//  GlobalConstants.swift
+//  ecommerce
+//
+//  Created by Vincent Salinas on 7/11/22.
+//
+
+import Foundation
+
+enum GlobalConstants: String {
+    case app = "Pick'n'Pay"
+    case aboutTitle = "BRINGING YOU THE BEST AUDIO GEAR"
+    case about = "Located at the heart of New York City, Amazona is the premier store for high end headphones, earphones, speakers, and audio accessories. We have a large showroom and luxury demonstration rooms available for you to browse and experience a wide range of our products. Stop by our store to meet some of the fantastic people who make Amazona the best place to buy your portable audio equipment."
+    case footer = "Amazona is an all in one stop to fulfill your audio needs. We're a small team of music lovers and sound specialists who are devoted to helping you get the most out of personal audio. Come and visit our demo facility - we’re open 7 days a week."
+    case copyright = "Copyright 2022. All Rights Reserved"
+}
+
+enum Categories: String {
+    case home = "HOME"
+    case headphones = "HEADPHONES"
+    case speakers = "SPEAKERS"
+    case earphones = "EARPHONES"
+}
+
