+++ ecommerce/ecommerce/Shared/About/About.swift
@@ -17,11 +17,11 @@ struct About: View {
                 .clipped()
                 .listRowInsets(EdgeInsets())
                 .cornerRadius(15)
-            Text("BRINGING YOU THE BEST AUDIO GEAR")
+            Text(GlobalConstants.aboutTitle.rawValue)
                 .font(.title)
                 .bold()
                 .multilineTextAlignment(.center)
-            Text("Located at the heart of New York City, Amazona is the premier store for high end headphones, earphones, speakers, and audio accessories. We have a large showroom and luxury demonstration rooms available for you to browse and experience a wide range of our products. Stop by our store to meet some of the fantastic people who make Amazona the best place to buy your portable audio equipment.")
+            Text(GlobalConstants.about.rawValue)
                 .multilineTextAlignment(.center)
         }
         .padding([.leading, .trailing], 27)
